BUG: Another piece of the fix for expanding mailing lists in recent Thunderbird versions: gMsgCompose.checkAndPopulateRecipients no longer exists -- it was replaced by gMsgCompose.expandMailingLists that we're now calling -- so don't try to call it anymore. This fix doesn't have any functional impact, other than eliminating an error message that was showing up for some users in the error console.
BUG: Fix a bug which was preventing a preference observer in the
message composition window from being removed properly when the window
was closed. This bug may have been a minor memory leak, but it's hard
to say for certain. This fix doesn't have any functional impact that I
know of, other than eliminating an error message that was showing up
for some users in the error console.