Another attempt to fix the error messages showing up in the console when the send later dialog window pops up.
Also, the attempt to fix this in the previous beta release broke some
functionality, in particular, when editing a previously scheduled
message and popping up the send later dialog, the "Send at" button
was greyed out until you edited the time field, when it should have
been active by default since the dialog already had a time filled in
(the time from when the message was previously scheduled). This is now
working again.