Fi accelerator keys Mac, change modifier on Linux and Windows
The "Alt" modifier which we were using for accelerator keys in the
prompt window doesn't do anything useful on the Mac, since it's bound
to the Option key, which can't be used without another modifier.

Therefore, change from "Alt" to "Accel", which is what we were
supposed to be using anyway, which binds the accelerator keys to the
Command key on the Mac. Note, however, that this changes the
accelerator modifier from Alt to Ctrl on Windows and Linux, which is
more standard.