BUG: Send Later stops working after system clock set in the future
If the user runs Thunderbird when the system clock is incorrectly set
in the future, and then subsequently the system clock is fixed, Send
Later will stop delivering scheduled messages.

This commit fixes that by detecting when the clock has moved backward
and resetting Send Later's internal state when that happens.