Disable scheduling messages when Enigmail is active
Unfortunately, limitations in the Thunderbird add-on framework make it
somewhere between difficult and impossible to make scheduled messages
with Send Later work safely and properly when the Enigmail add-on is
installed and enabled.

However, Send Mail can successfully deliver previously scheduled
messages even when Enigmail is active.

Therefore, this commit changes the add-on to display an alert pop-up
and disable the scheduling of messages when Enigmail is enabled.
If/when Enigmail is disabled or uninstalled and Thunderbird is
restarted, Send Later will automatically re-enable itself.

Addresses #1.