Add support for time restrictions
When scheduling a recurring message, the user can now specify start
and end times between which the message must be delivered. This is
mostly useful for minutely messages, but is also relevant to other
recurring messages when a message is not delivered at the scheduled
time, e.g., because Thunderbird was not running at that time. For that
case, there's a new preference indicating whether the user wants to
enforce the time restriction on the late message.

This commit also includes a bunch of code cleanup I did while
implementing the time restriction.