Make compose toolbar commands go through compose-send-message event
Make the buttons and pickers in the compose window toolbar go through
the compose-send-message event rather than scheduling messages
directly, so that they use the same workflow as the rest of the
compose window.

This is a step toward making Send Later compatible with Enigmail,
because to do that, we need to centralize the Send Later workflow so
that there are simple hooks that Enigmail can call to interact with
it.

As a bonus, this means that spell-checking before send and checking
for a missing Subject line will now happen as expected when using the
toolbar to schedule a message, whereas previously they were not.