Eliminate double compose-send-message on "Put in Outbox"
Refactor the message-scheduling code so that when the user clicks the
"Send Now" or "Put in Outbox" button in the dialog, it works by
falling through and allowing the rest of the standard message send
process to complete, rather than by re-creating it.

This eliminates a double call to the compose-send-message event when
doing "Put in Outbox", and simplifies the code flow in preparation for
adding back Enigmail support.

This has two other side effects.

The first is a bug fix: "Put in Outbox" will no longer cause
spell-checking to occur twice.

The second is that now the "Put in Outbox" button is only enabled when
the dialog was launched via Ctrl-Shift-Enter or File | Send Later, and
the "Send Now" button is only enabled when the dialog was launched by
clicking the Send button or Control Enter with the "Send Does Send
Later" preference enabled. This change is arguably a feature, since it
makes it harder to accidentally use one of these buttons when it
wasn't intended.