Stop overloading the meaning of the "sending_later" window attribute
Previously, we were using the "sending_later" attribute on the compose
window to mean one of two different things:

1) We've already initiated the Send Later process, and this is
actually the second time through the compose-send-message event
handler, being called from inside the GenericSendMessage that we
initiated.

2) The user has typed Ctrl-Shift-Enter, but the "alt_binding"
preference is enabled, so our compose-send-message event handler
shouldn't actually do anything.

This overloading of the attribute was confusing and interfering with
my ability to figure out how to make Enigmail work, so I've changed
the code to use two different attributes instead of one.