Remove rather than disable "Put in Outbox" and "Send Now" buttons
When we don't want the user to accidentally click the "Put in Outbox"
button because "Send" rather than "Send Later" was used to activate
the dialog, hide the button completely rather than just disabling it.

Similarly, when we don't want the user to accidentally click the "Send
Now" button because "Send Later" was used to activate the dialog, hide
it completely rather than just disabling it.