Add preference to always show "Send Now" and "Put in Outbox" buttons
Add a preference which, if enabled, causes both the "Send Now" and
"Put in Outbox" buttons to always be displayed in the Send Later
dialog, rather than displaying one or the other based on how the
dialog was opened.

The only-display-one-or-the-other functionality was added in Send
Later release 5.1 in an effort to minimize the incidents of people
using these buttons incorrectly. However, since the 5.1 release, I've
heard from a number of people whose workflows depended on both buttons
being present all the time. I had not anticipated these workflows when
making the change. Since they are reasonable workflows, I am adding
this preference to enable them.