Revert "Add preference to always show "Send Now" and "Put in Outbox" buttons"
Summary: If you need the "Put in Outbox" button in the Send Later dialog, then you need to launch the dialog using Ctrl-Shift-Enter, the "File | Send Later" menu command, or the button provided by the "Send Later Button" add-on. If, on the other hand, you need the "Send Now" button in the Send Later dialog, then you need to have the "'Send' does 'Send Later'" preference enabled, and you need to launch the dialog using the "Send" button, Ctrl-Enter, or the "File | Send Now" menu command. Please read on for the details surrounding this change.

There has been a lot of flip-flopping in this area of the add-on's functionality, caused in part by the fact that there are so many moving parts that I've had trouble keeping them all straight in my mind, so I'm going to try to explain clearly here exactly what's going on and why I'm reverting this change, so that all the necessary information is known going forward.

There are two buttons in the Send Later prompt window: "Put in Outbox" and "Send Now". These two buttons bypass the Send Later message scheduling mechanism in two different ways:

* "Put in Outbox" puts the message being edited directly into the Thunderbird Outbox folder, such that it will be delivered the next time the user (or Send Later, or BlunderDelay) executes the "File | Send Unsent Messages" menu command.

* "Send Now" delivers the message to its intended recipients immediately rather than scheduling the message for later delivery with Send Later.

Unfortunately, the only way that both of these buttons can always work is if Send Later "cheats" and does things it's not supposed to during the message scheduling and delivery process. This cheating causes all sorts of other issues, such as incompatibilities with other add-ons, Thunderbird popping up warnings and spell-checking windows multiple times, etc., so it's a rather bad idea, and we don't want Send Later to be doing that.

Therefore, in a recent release of Send Later, the cheating was removed, and that meant it was no longer possible for both of these buttons to work properly every time the Send Later dialog is displayed. In particular:

* If the Send Later dialog is opened with Thunderbird's own Send Later command, i.e., typing Ctrl-Shift Enter, selecting the "File | Send Later" menu command, or using the button provided by the "Send Later Button" add-on, then it is possible to make the "Put in Outbox" button work properly in the dialog, but it is _not_ possible to make the "Send Later" button work.

* Conversely, if the user has the "'Send' does 'Send Later'" preference enabled, such that Thunderbird's own "Send Now" command (i.e., typing Ctrl-Enter, selecting the "File | Send Now" menu command, or clicking the "Send" button) is used to launch the Send Later dialog, then in that particular instance of the dialog, it is possible to make the "Send Now" button work properly, but it is _not_ possible to make the "Put in Outbox" button work.

Because the removal of the cheating code meant it was no longer possible to make both of these buttons work every time the dialog was opened, Send Later's code was modified to display only the button that will actually work properly.

After the release with this change was shipped, I started to get feedback from users complaining -- in some cases rather vociferously -- about this change. Some people complained about the disappearance of the "Send Now" button, others complained about the missing "Put in Outbox" button.

At this point, I should have explained as above why it was no longer possible for both of these buttons to be present, and further explained that if the user needs the "Send Now" button, the dialog needs to be launched with the "Send" button or Ctrl-Enter or "File | Send Now", whereas if the "Put in Outbox" button is needed, the dialog needs to be launched with the "Send Later" button or Ctrl-Shift-Enter or "File | Send Later".

Unfortunately, with all of the complaints coming in, I got confused, and I somehow convinced myself that the buttons were removed not because it was not possible to make them work all the time without cheating, but rather because users were sometimes getting confused and clicking one of these buttons when it wasn't actually needed. Actually, reducing the likelihood of user confusion wasn't the reason why the buttons were removed; it was merely a beneficial side-effect.

I need to apologize to those users whom I told that the buttons were removed to reduce user confusion. I was wrong. I'm sorry for giving you incorrect information.

I also need to apologize to everyone for the "churn" of putting out release 5.2 which added a preference which puts both buttons back, when in fact with that preference enabled, _one of the buttons in the dialog is always going to function incorrectly_. So rather than solving the problem I hoped to solve in release 5.2, I introduced a new problem, an actual bug, which could cause messages to be put into the user's Outbox when the user intended to send them, or sent directly when the user intended to put them in the Outbox. I'm sorry.

Release 5.2, which introduced potentially incorrect behavior as described above, has been removed from addons.mozilla.org and is no longer available for downloads or automatic upgrades.

This new release undoes the change introduced in release 5.2. There is no longer a preference to cause both of the buttons to be displayed all the time, because it is simply not possible to make that work properly without Send Later interacting with Thunderbird in a way that it's not supposed to and that isn't sustainable or maintainable.

Just to reiterate, if you need the "Put in Outbox" button in the Send Later dialog, then you need to launch the dialog using Ctrl-Shift-Enter, the "File | Send Later" menu command, or the button provided by the "Send Later Button" add-on. If, on the other hand, you need the "Send Now" button in the Send Later dialog, then you need to have the "'Send' does 'Send Later'" preference enabled, and you need to launch the dialog using the "Send" button, Ctrl-Enter, or the "File | Send Now" menu command.

If you _don't_ have the "'Send' does 'Send Later'" preference enabled, and you launch the Send Later dialog the normal way (with Ctrl-Shift-Enter, "File | Send Later", or the "Send Later Button" button), and then you decide that actually you want to send the message right away, all you need to do is close the Send Later dialog or click the Cancel button and then send the message normally.

I am sorry to those of you whose workflows are disrupted by this change. If you are unable to figure out a new workflow which accommodates this change, _please_ email me (jik@kamens.us) and describe to me in detail how you're using the add-on and how this change is impacting you, and I will try to help you find a way to make it work.

This reverts commit 1b5c5d7d6d555bcf3a2f6fdcced5a1213b735a04.