Avoid using "instanceof" in JavaScript
As per http://perfectionkills.com/instanceof-considered-harmful-or-how-to-write-a-robust-isarray/,
don't try to use instance of to check object types in JavaScript,
because objects created within one iframe or context do not share
Prototypes with arrays created within other iframes or contexts. As a
result, if you create an object in a jsm and then try to use
instanceof on it elsewhere, it won't work. To avoid this issue, check
for methods that are expected on objects of specific types, e.g.,
"splice" for an Array or "getTime" for a Date.