Make Send Later compatible with the SpamFighter add-on
Disable the SpamFighter compose-send-message event listener, whose
only purpose is to sometimes add an advertising footer to the bottom
of outgoing messages, and which in the process of doing so breaks Send
Later and other add-ons.

If the makers of SpamFighter weren't so blasé about violating
Thunderbird add-on coding standards and doing stuff that breaks other
add-ons, then I might be reluctant to disable their event listener and
thereby prevent them from adding advertising to the bottom of users'
outgoing messages. But given that their code is crap that breaks the
rules in numerous ways, I have no qualms about this.