Eliminate the use of "eval" to parse dynamic scheduling arguments
Use JSON rather than eval to parse dynamic scheduling arguments.

Also, fix a bug in the interpretation of numeric return values (i.e.,
editor.

Also, fix a bug in how arguments returned by a function in the editor
were propagated into the arguments text box for the next invocation of
the function.

Also, fix a bug in how function results were displayed in the editor
so that everything returned by the function is displayed properly.