BUG: Fix incorrect invocations of Array.splice
There was some incorrect invocations of Array.splice which were
treating the second argument as the end index of the splice, rather
than as the number of elemented to replace in the array. These
incorrect invocations weren't actually having a functional impact
because of the way the code was otherwise constructed, but they could
at some point in the future and they were clearly wrong, so they've
been fixed.