BUG: Fix undefined variable error when throwing an internal error
There was an undefined variable reference in some code that should
never actually be executed since it indicates an internal error in the
add-on. However, I ran into that internal error during developement,
and then the exception I was expecting to be thrown wasn't because of
the bug in the exception-handling code itself. This is now fixed.