ENH: Improvements to day/time restrictions and scheduling functions
Previously, the day-of-week and time-of-day restrictions in the Send
Later prompt window could only be used with recurring messages or
scheduling functions. Now they can be used when specifying only a
single send time explicitly as well. You can do some clever things
when combining this with saving defaults in the prompt window. For
example, if you always want messages to be sent within a particular
time window, you could put "now" in the prompt window's text box and
specify the desired time window in the "Between" settings, then save
those defaults, and from then on, whenever you bring up the prompt
window and click the "Send at" button, it'll either send the message
immediately -- if it's within the correct time window -- or move it to
be within the time window.

Previously, when you used a non-recurring dynamic scheduling function
such as the "BusinessHours" function that is created by default when
you open the dynamic function editor, the fact that the message was
scheduled with a function is not saved, so if you go to edit the
function later and then try to re-schedule it, rather than defaulting
to using the same dynamic function again, it defaulted to the time
that the message was previously scheduled to be sent. Now, the
function name is saved, and if you edit and re-schedule the message
the prompt window will be pre-configured to use it again.

Furthermore, when you specify day-of-week or time-of-day restrictions
when scheduling a message with either an explicit send time or a
non-recurring dynamic function, the restrictions you specify will now
be saved, so that if you edit and re-schedule the message, they will
be set up in the prompt window automatically.