BUG: Fix incorrect substitution strings in German, Hebrew, Swedish
Send Later's translations use parameterized substitutions in some
translated strings, i.e., the string is used as a template into which
values are subsituted to customize it for the particular message being
displayed. A few of the rarely used strings had missing or incorrect
substitution patterns in them. This has been fixed.

In addition, the build process for the add-on has been modified to
detect these issues in the future.