Fix "Close" command key and access key in function editor
The code for the dynamic function editor was incorrectly using a
single key setting for both the command key, i.e., the ctrl-whatever
key combination, and the access key, i.e., the Alt-whatever key
corresponding to one of the letters in the button label, for the Close
button.

The command key should always be ctrl-w, the standard "close window"
command regardless of language. The code has now been fixed to reflect
this.

The access key, on the other hand, should be the earliest available
letter in the localized label for the Close button. This, too, is now
fixed.

In addition, the "Save" access key for Finnish, French, and Italian
have been fixed to match their button labels.