BLD: Message about replacing strings should be debug, not warning
Nowadays I'm shipping with incomplete translations on a regular basis,
so I expect there to be replaced strings every time I build the
add-on. Therefore, don't display the messages about these strings in
the build output, because they make it harder to see the actual errors
I need to worry about.