Localization: Put add-on name in just one place, sub it elsewhere
To make it easier to be consistent about referring to the name of the
add-on, the contents of the "MessageTag" entry in the
background.properties localization file is now considered the
canonical name of the add-on in any given translation. Elsewhere in
translations when the add-on name is used, it is now referred to as
{NAME}, which is substituted automatically with the name of the add-on
when the add-on is built.

I had to delete some translated strings while making this change,
because they were in foreign, non-Latin languages I couldn't interpret
well enough to figure out where the name of the add-on was so that I
could replace it with {NAME}. These strings will have to be
re-translated.