Automatically check locale data in install.rdf and chrome.manifest
Make sure all locale directories are listed in install.rdf and
chrome.manifest. Make sure there are no extra locales listed in
install.rdf or chrome.manifest. Make sure the add-on names and
descriptions in install.rdf match the ones in the locales.

This turned up a bunch of inconsistencies, which have now been fixed.