Compact Outbox after sending messages
Thunderbird's local folder storage format keeps deleted messages on
disk, but invisible to the user, until their folder is compacted.

Usually this causes no trouble, but occasionally Thunderbird corrupts
a local folder and loses track of messages which messages were
deleted, causing them to "magically" appear in the folder again.

In most cases, this causes no further harm than deleted messages being
resurrected, an annoying nuisance but not a huge problem.

However, when Send Later is being used, and when the folder this
happens to is the user's Outbox folder, it can cause a _big_ problem:
messages that were previously sent with Send Later can be resurrected
and resent. If it has been a long time since the Outbox was last
compacted, a _lot_ of messages can be resent. This is bad.

I can't fix the root cause of this issue, i.e., whatever causes
folders to occasionally be corrupted like this, because that happens
inside Thunderbird, not inside Send Later. However, this change makes
the problem much, much less likely to occur.

It does this by automatically compacting the Outbox folder whenever
Thunderbird finishes sending unsent messages (i.e., when Send Later
messages are finished being delivered) and the Outbox is visibly
empty, i.e., the only messages remaining in it are invisible, deleted
messages.

No, Send Later shouldn't need to do this, but it is what it is.