Make locale-based first-day-of-week work in calendar pop-ups
There is a bug in Thunderbird which was preventing the calendar
pop-ups displayed by Send Later from starting the week on the correct
day, in locales where that day isn't Sunday. See
https://bugzilla.mozilla.org/show_bug.cgi?id=516796. Fortunately,
there's a workaround for this bug, which has now been implemented in
Send Later. Therefore, if your computer starts weeks with the correct
day elsewhere, it should now do so in the Send Later calendar pop-ups
as well.