Fix locale_import rule because list of locale files changes
Apparently Make's $(shell ...) expansions inside a rule happen at the
beginning of the rule, not as each command is executed. This means
that if the files matching the expansion change while the rule is
executing, the changes won't be picked up by subsequent commands in
the rule. Therefore, change the locale_import rule to use shell
substitutions directly in commands rather than Make's $(shell ...)
substitutions.