BUG: "cancel recurrence on reply" getting set incorrectly sometimes
Some recurring messages were ending up with their "cancel recurrence
on reply" flags set when they shouldn't have been.

I'm not exactly sure of the circumstances leading to this, but I think
at least one of them was when using user-defined functions assigned to
toolbar buttons.

The underlying issue leading to this behavior was that although the
empty string "" is interpreted as false by JavaScript, an empty string
_object_, e.g., `new String("")`, isn't, so it's a bad idea to rely on
empty string evaluating to false. Gotta love JavaScript!