Add hidden preference to set first day of week in date pickers
The date-picker grid, i.e., the calendar that pops up to let you click
on a date to select it, figures out what to use as the first day of
the week by looking at the locale Thunderbird is running in.

It does _not_ look at the user's system settings, so if the user has
told the operating system to use a different first day of the week,
the date-picker won't reflect that.

Unfortunately, it is impossible (as far as I know) for Send Later's
code running inside Thunderbird to query the operating system to ask
what the user has set as the first day of the week, so I can't
completely fix this bug. What I can do, however, is add a
configuration setting to allow the user to tell Send Later what to use
as the first day of the week.

Because this setting is of interest to very few people, I've decided
not to add it to the Send Later preferences window. Instead, it is a
hidden preference called "extensions.sendlater3.first_day_of_week"
which you need to use the advanced configuration editor to modify. If
its value is -1, the default, then the locale default will be used.
Otherwise, if it's value is between 0 and 6, then the specified day (0
is Sunday) will be used as the first day of the week in date-picker
grids.