BUG: Changes to time picker not detected when toolbar button clicked
Old behavior:

If you click in the time picker in the compose window toolbar, edit
the hour or minute value, and then click the Send Later button in the
toolbar while the time picker is still focused, then the change you
just made to the value of the time picker is lost and the message is
scheduled with the previous hour or minute value.

New behavior:

When you click the Send Later button in the toolbar, the changes you
just made to the time picker values are noticed and used to schedule
the message.