Load our default preferences; don't rely on TB to do it
The code to load add-ons' default preferences was removed from
mozilla-central, and therefore from Thunderbird. They're feverishly
working to put it back, but I'm not sure the workaround they're
implementing is going to stay around forever, and in any case we won't
be able to rely on that anymore once we convert the add-on to be
bootstrapped, and furthermore I need to be able to test Send Later on
interim builds where they haven't yet fixed the problem. Therefore,
I'm adding code to Send Later to explicitly load the default
preferences rather than relying on the app to do it.