XUL datepicker/timepicker Thunderbird compatibility changes
The datepicker and timepicker XUL widgets used by Send Later are
moving from the mozilla-central source tree (i.e., the source three
shared by Firefox, Thunderbird, and SeaMonkey) into comm-central,
i.e., the source tree used only by Thunderbird. Changes to Send Later
are needed to accommodate this change.

In addition, a bug fix to the timepicker that I introduced into Send
Later by making my own copy of the timepicker source code has been
merged into the Thunderbird source code as of Thunderbird 61
(https://bugzilla.mozilla.org/show_bug.cgi?id=1389791).

The following changes need to be made to Send Later to accommodate all
of the above:

* Starting in Thunderbird 61, use the datepicker and timepicker
  widgets that ship with Thunderbird rather than my custom versions.
* For all SeaMonkey versions and Thunderbird versions prior to 61, use
  my custom datepicker and timepicker bindings (previously, my custom
  timepicker binding was being used but not datepicker). This might
  not be entirely necessary, i.e., there are probably still some app
  versions where we could use the built-in datepicker widget, but
  using ours simplifies things.
* In all of my XUL files, include the global Thunderbird XUL bindings
  stylesheet as well as the global mozilla-central XUL bindings
  stylesheet (necessary not only for datepicker and timepicker, but
  also for any other mozilla-central XUL widgets that are moved from
  there into comm-central later).
* Store the default first-day-of-week value for our custom datepicker
  widget in our localization rather than relying on the application to
  provide it. This is simpler than conditionalizing our datepicker
  with a separate path to the stylesheet locale stylesheet for each
  application.

Note that I'm still waiting to hear back on some additional
information to help me figure out how SeaMonkey is going to play out
in all this, but the changes in this commit should be correct moving
forward for Thunderbird.

Note also that some of the stylesheets I'm including in XUL files may
not be available everywhere the add-on runs, but this is apparently
safe because apparently missing stylesheets are simply ignored rather
than causing interpretation of the XUL to abort.