Patch SugarJS for compatibility with Thunderbird 61+
In Thunderbird 60 and earlier, when SugarJS is loaded, the global
context it identifies is, indeed, a global context with access to all
global variables and types (Array, Date, etc.). In particular, it's a
`BackstagePass` object.

However, starting in TB 61, it's a `NonSyntacticVariablesObject` which
does _not_ have access to global variables and types, so it fails to
load properly.

This commit adds a different method for SugarJS to use when loading to
locate the global context, one which works successfully on TB 61+. For
paranoia, this change only uses the new method when the other method
fails to find the right global context, though I think it could
actually work all the time.