Fix bug in date-picker for some older Thunderbird versions
The API in Thunderbird for formatting dates has been in a great deal
of flux for the past several Thunderbird releases. Release 6.4.1 of
Send Later was an attempt to make the date code in Send Later
compatible up to Thunderbird 60.3.0. Unfortunately, although the
changes did achieve compatibility with newer Thunderbird releases,
they broke compatibility with older ones. This release attempts to
address that.