Fix UTF-8 dates in Send button in prompt window
The label for the button in the prompt window that displays the
scheduled send time after the user enters a valid date was being
garbled for locales when dates use non-ASCII characters. This was
occurring because the SugarJS module containing the code for
formatting those dates was being loaded ASCII instead of UTF-8, so the
UTF-8 strings in that code were being corrupted. The module is now
being loaded correctly, so those dates are displayed correctly.