Restore date-picker and time-pickers to Send Later prompt window
For Thunderbird 68+, restore the date-picker and time-pickers to the
Send Later prompt window, which means that it is once again possible
to schedule a message exclusively using the mouse, and it is once
again possible to put time-of-day restrictions on the delivery of
scheduled messages.

Two caveats:

1. The date-picker and time-pickers are only available in Send Later
   when the Lightning add-on is installed and enabled. This is because
   there is still no working date-picker or time-picker functionality
   in the core Thunderbird application, so Send Later is for the time
   being using the date-picker and time-picker included with
   Lightning.

2. The Send later date/time-picker that could previously be added to
   the message composition toolbar is still unavailable, because the
   date-picker and time-picker in Lightning doesn't work in toolbar
   buttons.

This change adds one additional new bit of functionality that wasn't
there before: when you change either the start or end time in the
"Between" range in the prompt window, the other time bracketing the
range is adjusted to preserve the validity of the range. So, for
example, if the start time is set to 11:00 AM and then you change the
end time to 10:00 AM, the start time will change to 10:00 AM
automatically.