Reorganize some locale strings to accommodate Hungarian translations
The day of the week is written differently in Hungarian depending on
whether it stands alone, has the word "on" before it in English (in
which case the "on" is incorporated into the day of the week itself in
Hungarian), or is in the phrase "nth <weekday> of the month".
Previously we used one set of "day-of-the-week" translation strings
for all of these purposes. to accommodate Hungarian, we now need three
separate sets, which will be identical in many languages but not in
Hungarian.