Fix "double down arrows" in compose buttons
Send Later was causing the buttons in the compose formatting toolbars
to have two down arrows in them. This appears to be because we were
including chrome://global/skin/global.css in one of our overlays when
that stylesheet is already included in the compose window by
Thunderbird itself so we don't need to include it as well.

h/t Óvári for the bug report.