Get rid of "<from locale>" in preferences window
Rather than showing "<from locale>" for each shortcut button label
text when the actual label for that button is being pulled from Send
Later's localization for the current language, show the actual value
of the localization. This removes a lot of unnecessary confusion about
what "<from locale>" means.

Thanks to Óvári for suggesting that I localize the "<from locale>"
string, a reasonable suggestion but not actually the right fix, which
is to get rid of that string entirely, something that others have
suggested in the past.