Fix bug in default preferences loading code
My understanding of how JavaScript destructuring assignment works was
incorrect, and as a result, my code for loading default preferences
was also incorrect. It was wrapped in a try/catch for pre-TB-68
backward compatibility, so the try/catch was hiding the error in the
code. The try/catch has been removed and the code has been fixed.

I'm not actually sure this has any functional change, because I think
current TB's legacy extension loader loads default preferences
loading, but I'm not 100% certain about that, so I want to make sure
the code is correct.