Fix compose toolbar buttons for Thunderbird 71
FIX: A change to the core Thunderbird application after Thunderbird 68
broke Send Later's composition window toolbar buttons. This is now
fixed.

Technical details:

Now that the toolbar has been converted into a custom element, we
can't rely on it being created and populated when the the load event
for the window is called. Therefore, we now wait to manipulate the
buttons until after the NotifyComposeBodyReady event.

In addition, there's non-functional prep work in this commit to
prepare for being able to disable the Ctrl-Alt-1, 2, and 3 composition
window "hot-keys": these keys are now being created in JavaScript
rather than in the XUL file. This way we can not create them if the
user has disabled them in their preferences. That change, however, has
not yet been implemented.