Fix displaying / hiding of "Send Later" message list column in TB68
There appears to be a bug in Thunderbird 68 which causes the "Send
Later" column in the message list to be displayed sometimes when it
shouldn't be. I've been able to reproduce one of the manifestations of
this bug reported to me, but another. This fix addresses the one I was
able to reproduce, and I'm hoping it will also fix the other one,
though I'm not certain.