Support a simple n-minute delay for all sent messages!
This commit adds a new feature which many people have asked for over
the years: a simple way to configure Send Later to automatically delay
all sent messages by a fixed number of minutes, i.e., an easy
replacement for the old "BlunderDelay" extension.

In addition to the "'Send' does 'Send Later'" preference, there is a
new, mutually exclusive "'Send' delays messages by # minutes". When
this preference is enabled, then whenever a user sends a message with
the Send button or Ctrl-Enter, it will be scheduled by Send Later to
be sent after the specified number of minutes.