textbox -> html:input
For compatibility with Thunderbird 71+, use html:input instead of textbox.

Note: "flex" doesn't work on html:input elements and I can't figure
out how to make it work, so the text input fields in the dynamic
function editor which were previously neatly aligned with the right
edge of the window aren't anymore. D'oh.