Fix potential message corruption in non-English locales
In some foreign languages it was possible for scheduled messages to be
corrupted when they were sent, because a date being inserted at the
top of the message contained a textual description of the sender's
time zone, which could include foreign-language characters which
confused Thunderbird into thinking there was a line break in the
message where there shouldn't be one.

This has been fixed by no longer including the textual time zone
description into the message.