ENH: Accept CRLF line breaks in dynamic function export files
Now that dynamic function export files, i.e. ".slj" files, have line
breaks in them, it's possible for them to be transmitted and copied in
such a way that their LF line terminators get turned into CRLF, e.g.,
if someone views a .slj file in a browser and then copies and pastes
it into a text file on Windows. There's no reason why we can't support
this, so let's do it.