Add the Kickstarter campaign pop-up and disable donation pop-up
I can no longer afford to maintain my add-ons for for free, so I am
launching a Kickstarter campaign to try to get enough users to commit
to supporting them so that I can afford to keep maintaining them. I'm
adding pop-ups to all of my add-ons letting people know about this and
asking them to participate in the Kickstarter campaign.

Because the add-on is going to be licensed for a fee starting in TB78,
I'm disabling the donation pop-up.